+++ src/main/java/io/github/profetgit/tidypockets/mixin/CreativeScreenInvoker.java
@@ -1,0 +1,12 @@
+package io.github.profetgit.tidypockets.mixin;
+
+import net.minecraft.client.gui.screens.inventory.CreativeModeInventoryScreen;
+import net.minecraft.world.item.CreativeModeTab;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Invoker;
+
+@Mixin(CreativeModeInventoryScreen.class)
+public interface CreativeScreenInvoker {
+    @Invoker("selectTab")
+    void tidypockets$selectTab(CreativeModeTab tab);
+}